# openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Karag API",
    "description": "Modular RAG & Agentic Chatbot API",
    "contact": {
      "name": "Karag Team",
      "url": "https://github.com/karag-ai/karag"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/workspaces": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "List Workspaces",
        "operationId": "list_workspaces_api_v1_workspaces_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_list_Workspace__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Create Workspace",
        "operationId": "create_workspace_api_v1_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_Workspace_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/workspaces/": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "List Workspaces",
        "operationId": "list_workspaces_api_v1_workspaces__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_list_Workspace__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Create Workspace",
        "operationId": "create_workspace_api_v1_workspaces__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_Workspace_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}": {
      "patch": {
        "tags": [
          "workspaces"
        ],
        "summary": "Update Workspace",
        "operationId": "update_workspace_api_v1_workspaces__workspace_id__patch",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_Workspace_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workspaces"
        ],
        "summary": "Delete Workspace",
        "operationId": "delete_workspace_api_v1_workspaces__workspace_id__delete",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "name": "dataset_delete",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Dataset Delete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/details": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Get Workspace Details",
        "operationId": "get_workspace_details_api_v1_workspaces__workspace_id__details_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_WorkspaceDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/share-document": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Share Document",
        "operationId": "share_document_api_v1_workspaces__workspace_id__share_document_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareDocumentPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/graph": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Get Workspace Graph",
        "operationId": "get_workspace_graph_api_v1_workspaces__workspace_id__graph_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppResponse_WorkspaceDetail_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceDetail"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[WorkspaceDetail]"
      },
...
      "AppResponse_Workspace_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Workspace"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[Workspace]"
      },
...
      "AppResponse_dict_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[dict]"
      },
...
      "AppResponse_list_Workspace__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Workspace"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[list[Workspace]]"
      },
...
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique document ID (stable, persistent)"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "description": "Workspace this document belongs to"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Original filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "description": "MIME type"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Source URL or description"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Processing status",
            "default": "uploaded"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size",
            "description": "File size in bytes"
          },
          "chunks": {
            "type": "integer",
            "title": "Chunks",
            "description": "Number of chunks indexed",
            "default": 0
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Last update timestamp"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "Document content (truncated for preview)"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url",
            "description": "Presigned URL for download"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "filename",
          "content_type"
        ],
        "title": "DocumentResponse",
        "description": "Standard document response for frontend integration.\n\nThis schema is used for:\n- Document listing in workspace storage\n- Citation click-through from chat\n- Document metadata display",
        "example": {
          "chunks": 42,
          "content_type": "application/pdf",
          "created_at": "2024-01-15T10:30:00Z",
          "filename": "document.pdf",
          "id": "doc_abc123",
          "size": 1024000,
          "source": "upload",
          "status": "indexed",
          "updated_at": "2024-01-15T10:35:00Z",
          "workspace_id": "ws_xyz789"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ShareDocumentPayload": {
        "properties": {
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "target_workspace_id": {
            "type": "string",
            "title": "Target Workspace Id"
          }
        },
        "type": "object",
        "required": [
          "source_name",
          "target_workspace_id"
        ],
        "title": "ShareDocumentPayload"
      },
...
      "ThreadMetadata": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "default": "New chat"
          },
          "has_thinking": {
            "type": "boolean",
            "title": "Has Thinking",
            "default": false
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "workspace_id"
        ],
        "title": "ThreadMetadata"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Workspace": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Short unique ID of the workspace"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name of the workspace"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Detailed description"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceStats"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "Workspace"
      },
...
      "WorkspaceCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^[\\w\\s.-]+$",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "embedding": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding"
          },
          "retrieval": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval"
          },
          "generation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation"
          },
          "chunking": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunking"
          },
          "rag_engine": {
            "type": "string",
            "enum": [
              "basic",
              "graph"
            ],
            "title": "Rag Engine",
            "default": "basic"
          },
          "neo4j_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neo4J Uri"
          },
          "neo4j_user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neo4J User"
          },
          "neo4j_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neo4J Password"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "WorkspaceCreate"
      },
      "WorkspaceDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Short unique ID of the workspace"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name of the workspace"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Detailed description"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "threads": {
            "items": {
              "$ref": "#/components/schemas/ThreadMetadata"
            },
            "type": "array",
            "title": "Threads",
            "default": []
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentResponse"
            },
            "type": "array",
            "title": "Documents",
            "default": []
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "WorkspaceDetail"
      },
      "WorkspaceStats": {
        "properties": {
          "thread_count": {
            "type": "integer",
            "title": "Thread Count",
            "default": 0
          },
          "doc_count": {
            "type": "integer",
            "title": "Doc Count",
            "default": 0
          }
        },
        "type": "object",
        "title": "WorkspaceStats"
      },
      "WorkspaceUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "WorkspaceUpdate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}